tests/integration/requirements.txt
Flask
google-cloud
google-cloud-logging
google-cloud-monitoring
requests
